+++ database/migrations/2026_01_16_010050_create_call_sessions_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -19,8 +18,8 @@ return new class extends Migration
             $table->timestamp('started_at')->nullable();
             $table->timestamp('ended_at')->nullable();
             $table->unsignedInteger('duration_seconds')->nullable();
-            $table->unsignedBigInteger('rate_per_minute')->default(0);
-            $table->unsignedBigInteger('cost')->default(0);
+            $table->decimal('rate_per_minute', 8, 2)->default(0);
+            $table->decimal('cost', 10, 2)->default(0);
             $table->string('callerdesk_call_id', 191)->nullable();
             $table->json('meta')->nullable();
             $table->timestamps();
